Add VoiceChannelJoinMono
